backend/requirements.txt
Flask==3.0.*
Flask-SQLAlchemy==3.1.*
Flask-Migrate==4.0.*
Flask-CORS==4.0.*
requests==2.31.*
python-dotenv==1.0.*
Pillow==10.0.*
numpy==1.24.*
pandas==2.0.*
